refactor mapper
