+++ src/main/java/ua/everybuy/buisnesslogic/service/advertisement/AdvertisementCreationService.java
@@ -10,7 +10,7 @@ import ua.everybuy.database.entity.Advertisement;
 import ua.everybuy.database.entity.AdvertisementPhoto;
 import ua.everybuy.database.entity.AdvertisementStatistics;
 import ua.everybuy.routing.dto.mapper.AdvertisementResponseMapper;
-import ua.everybuy.routing.dto.mapper.AdvertisementToEntityMapper;
+import ua.everybuy.routing.mapper.AdvertisementToEntityMapper;
 import ua.everybuy.routing.dto.request.CreateAdvertisementRequest;
 import ua.everybuy.routing.dto.response.CreateAdvertisementResponse;
 import ua.everybuy.routing.dto.response.StatusResponse;
