+++ src/main/java/ua/everybuy/buisnesslogic/service/advertisement/AdvertisementUpdateService.java
@@ -5,12 +5,11 @@ import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
-import ua.everybuy.buisnesslogic.service.location.CityService;
 import ua.everybuy.buisnesslogic.service.photo.PhotoService;
 import ua.everybuy.database.entity.Advertisement;
 import ua.everybuy.database.entity.AdvertisementPhoto;
 import ua.everybuy.routing.dto.mapper.AdvertisementResponseMapper;
-import ua.everybuy.routing.dto.mapper.AdvertisementToEntityMapper;
+import ua.everybuy.routing.mapper.AdvertisementToEntityMapper;
 import ua.everybuy.routing.dto.request.UpdateAdvertisementRequest;
 import ua.everybuy.routing.dto.response.StatusResponse;
 import ua.everybuy.routing.dto.response.UpdateAdvertisementResponse;
