+++ src/main/java/ua/everybuy/routing/mapper/FavouriteAdvertisementMapper.java
@@ -1,10 +1,10 @@
-package ua.everybuy.routing.dto.mapper;
+package ua.everybuy.routing.mapper;
 
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
 import ua.everybuy.database.entity.Advertisement;
 import ua.everybuy.database.entity.FavouriteAdvertisement;
-import ua.everybuy.routing.dto.mapper.helper.CityMappingHelper;
+import ua.everybuy.routing.mapper.helper.CityMappingHelper;
 import ua.everybuy.routing.dto.response.AddToFavouriteResponse;
 import ua.everybuy.routing.dto.response.FavouriteAdvertisementResponse;
